Rename LongTracking for backward compatability. Fix an error in SimpleLongTracking. Add logic to use SimpleLongTracking for RelativeLoadBalancerStrategy.
